feat(quality): Implement Style Patrol feature for code quality analysis and refresh functionality
