+++ app/views/solidstats/quality/style_patrol.html.erb
@@ -3,7 +3,7 @@
   <div class="breadcrumbs text-sm mb-6">
     <ul>
       <li><a class="text-primary hover:text-primary-focus">Dashboard</a></li>
-      <li><a class="text-primary hover:text-primary-focus">Security</a></li>
+      <li><a class="text-primary hover:text-primary-focus">Quality</a></li>
       <li class="text-base-content/70">Style Patrol</li>
     </ul>
   </div>
@@ -30,7 +30,7 @@
           Ruby Standard gem code quality enforcement
         </p>
         
-        <%= link_to refresh_securities_style_patrol_path, method: :post, 
+        <%= link_to refresh_quality_style_patrol_path, method: :post, 
             class: "btn btn-accent btn-lg gap-3 shadow-xl hover:shadow-2xl hover:scale-105 transition-all duration-300 group",
             data: { turbo_submits_with: "Analyzing..." } do %>
           <svg class="w-5 h-5 group-hover:rotate-180 transition-transform duration-500" fill="none" stroke="currentColor" viewBox="0 0 24 24">
@@ -438,7 +438,7 @@
       </div>
       <ul tabindex="0" class="dropdown-content z-[1] menu p-2 shadow-2xl bg-base-100 rounded-box w-52 mb-2">
         <li>
-          <%= link_to refresh_securities_style_patrol_path, method: :post, class: "gap-3" do %>
+          <%= link_to refresh_quality_style_patrol_path, method: :post, class: "gap-3" do %>
             <svg class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
               <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15"></path>
             </svg>
